UpdSocket set to nonblocking mode
To have multiple nodes running on the same host discovery must bind to
the same port multiple times. Tokio did not expose the needed
SO_REUSEPORT. To work around this I used the socket2 crate which exposes
a safe interface to SO_REUSEPORT. The socket from that crate was then
tranformed to a std::net::udpsocket which was transformed to a
tokio::net::udpsocket.

When creating a new socket tokio sets nonblocking to true. It (no
longer) does this when creating a tokio socket from an std socket.

related tokio issue: https://github.com/tokio-rs/tokio/issues/3832