add instance id to tracing, mitigate warnings
It seems tracing spans are not passed to newly spawned `tokio tasks`. I
was using the outer span to make sure the `instance id` was recoverable
for each log msg. Now a new span with the `instance id` as field is
created whenever a tasks is spawned. The id is now part of the `State`
struct to make it easily availible in more places.