Heartbeats timeout now constant, elections improved
Constant Heartbeats
In the Raft paper the heartbeat timeout varies in length to prevent all
nodes simultaniously holding an election. In the design I use the raft heartbeat
duration to guarentee file lease validity.

[[
Even if a node goes down directly after issuing a lease it will take at
least HeartBeat T before a new lease can be issued, thus leases are
valid for at least HeartBeat T after issuing)
]]

This commit splits the classic heartbeat timing out leading to an election out
into 2 phases.

 - the heartbeat times out
 - a (bounded) random wait period
 - if not leader appeared the election is held

Elections Improved
Getting a majority of votes takes longer if a large part of the cluster
is down. This results in a lot of stalemates. To improve this the wait
period introduced above is increased for each consecutive failed election