Fixes clerk's log corrupting switching ministry
When a clerk switches to another ministry its log needs to be cleared.
However, as the minister of the clerks new ministry might receive the
reassignment order first it might already start sending raft entries
over.

This commit makes it so any reassignment changes the ministers' term to
be the highest of the cluster. Therefore, any messages it sends before
the clerk has switched will be rejected by the clerk as prev_log_term does
not match.

prev_log_term does not match because every ministries' term is now unique
(the highest is unique).

To bolt this onto raft request_gen I introduce a trait GetTerm and two
implementors FixedTerm and AtomicTerm. The first allowes no term
changing and is used by the president. The second allowes the term to be
changed using shared memory and is used by the minister now.